fivetran_connector/requirements.txt
fivetran-connector-sdk
paho-mqtt